edited supply pane

+++ src/main/scala/poolmate/pane/SupplyPane.scala
@@ -41,7 +41,7 @@ class SupplyPane(context: Context, model: Model) extends VBox:
           _.value.costProperty
         }
     )
-    prefHeight = context.height.toDouble
+    prefHeight = context.height
     items = model.supplyList
 
   supplyTableView.selectionModel().selectionModeProperty.value = SelectionMode.Single
